Working on integrating allstreamer's package versioning module
