+++ Cargo.toml
@@ -24,5 +24,8 @@ structopt = { version = "0.3.26", features = ["color"] }
 strum = "0.24.1"
 strum_macros = "0.24.2"
 indicatif = "0.17.0"
+lazy_static = "1.4.0"
+serde = {version = "1.0.143", features = ["derive"]}
+pomsky-macro = "0.5.0" # 0.6.0 doesnt work for me ;-;
 
 [dev-dependencies]
